# api/content-msv/v1/http/openapi.yaml
openapi: 3.0.0
info:
  title: Orpheon Content API
  version: 1.0.0
servers:
  - url: https://orpheon.space/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development
paths:
...
  /licenses:
    get:
      summary: Get all licenses
      tags:
        - licenses
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/License'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    post:
      summary: Create a new license
      description: Create a new music license with the provided information
      tags:
        - licenses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - type
              properties:
                name:
                  type: string
                  maxLength: 100
                  description: License name
                description:
                  type: string
                  maxLength: 1000
                  description: License description
                url:
                  type: string
                  description: license url
      responses:
        '201':
          description: License created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "License created successfully"
        '400':
          description: Bad request - invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /licenses/{id}:
    get:
      summary: Get license by ID
      description: Retrieve license details by its unique identifier
      tags:
        - licenses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/License'
        '400':
          description: Bad request - invalid license ID format
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: License not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    put:
      summary: Update license information
      tags:
        - licenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - type
              properties:
                name:
                  type: string
                  maxLength: 100
                  description: License name
                description:
                  type: string
                  maxLength: 1000
                  description: License description
                url:
                  type: string
                  description: license url
      responses:
        '200':
          description: License updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "License updated successfully"
        '400':
          description: Bad request - invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: License not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      operationId: updateLicense

    delete:
      summary: Delete license
      description: Delete a license by its unique identifier
      tags:
        - licenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: License unique identifier
      responses:
        '200':
          description: License deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "License deleted successfully"
        '400':
          description: Bad request - invalid license ID format
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

# Me
...
components:
  schemas:
...
    License:
      type: object
      required:
        - id
        - title
        - description
        - url
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 100
          description: License title
          example: "Creative Commons Attribution"
        description:
          type: string
          maxLength: 500
          description: License description
          example: "Allows redistribution and reuse with attribution"
        url:
          type: string
          format: uri
          maxLength: 500
          description: URL to license details
          example: "https://creativecommons.org/licenses/by/4.0/"
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        title: "Creative Commons Attribution"
        description: "Allows redistribution and reuse with attribution"
        url: "https://creativecommons.org/licenses/by/4.0/"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT